+++ src/tools/log.rs
@@ -68,8 +68,8 @@ impl LogObj {
             last_logging: Instant::now(),
         }
     }
-    pub fn update(&mut self, size: usize, samples: usize, srate: u32) {
-        self.total_size += size as u128;
+    pub fn update(&mut self, size: u128, samples: usize, srate: u32) {
+        self.total_size += size;
         self.duration.insert(srate, if self.duration.contains_key(&srate) { self.duration[&srate] } else { 0 } + samples as u128);
         self.bitrate.insert(srate, if self.bitrate.contains_key(&srate) { self.bitrate[&srate] } else { 0 } + size as u128);
     }
